+++ CHANGELOG.md
@@ -10,6 +10,17 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
+## [0.7.5] - 2024-08-17
+
+### Added
+
+- Added `MustMarshalCommandIntoEnvelope()`, `MustMarshalEventIntoEnvelope()` and `MustMarshalTimeoutIntoEnvelope()`.
+- Added `UnmarshalCommandFromEnvelope()`, `UnmarshalEventFromEnvelope()` and `UnmarshalTimeoutFromEnvelope()`.
+
+### Deprecated
+
+- Deprecated `MustMarshalMessageIntoEnvelope()` and `UnmarshalMessageFromEnvelope()`.
+
 ## [0.7.4] - 2024-06-21
 
 ### Added
@@ -153,6 +164,7 @@ different marshalers/unmarshalers provided by `google.golang.org/protobuf`.
 [0.7.0]: https://github.com/dogmatiq/marshalkit/releases/tag/v0.7.0
 [0.7.1]: https://github.com/dogmatiq/marshalkit/releases/tag/v0.7.1
 [0.7.3]: https://github.com/dogmatiq/marshalkit/releases/tag/v0.7.3
+[0.7.5]: https://github.com/dogmatiq/marshalkit/releases/tag/v0.7.5
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD